Adds net-zero-module-2 to repository.

+++ course-ul.sh
@@ -17,16 +17,22 @@ for f in *.zip; do
     read -p "Enter the name of this resource: " user_input
 
     folder_name=$(echo "${user_input// /-}" | tr '[:upper:]' '[:lower:]')
+    echo $folder_name
 
-    unzip -q $f content/* -d $folder_name; mv $folder_name/content/* $folder_name/; rm -r $folder_name/content
+    if [ ! -d $folder_name ]; then
+      unzip -q $f content/* -d $folder_name; mv $folder_name/content/* $folder_name/; rm -r $folder_name/content
 
-    # Adds noindex / nofollow to <meta> tag in <head> of index.html.
-    sed '/<head>/a<meta name='robots' content='noindex,nofollow'>'  $folder_name/index.html >  $folder_name/temp.html && mv  $folder_name/temp.html  $folder_name/index.html
-    
-    echo "Finished unzipping ${f}. Deploying to Github now." 
-    git add --all; git commit -m "Adds $folder_name to repository."; git push origin main
-    echo "Finished uploading. Use https://bc-it-services.github.io/etutorial/${folder_name}/index.html for the iFrame."
+      # Adds noindex / nofollow to <meta> tag in <head> of index.html.
+      sed '/<head>/a<meta name='robots' content='noindex,nofollow'>'  $folder_name/index.html >  $folder_name/temp.html && mv  $folder_name/temp.html  $folder_name/index.html
+      
+      echo "Finished unzipping ${f}. Deploying to Github now." 
+      git add --all; git commit -m "Adds $folder_name to repository."; git push origin main
+      echo "Finished uploading. Use https://bc-it-services.github.io/etutorial/${folder_name}/index.html for the iFrame."
 
+      rm $f      
+    else
+      echo "This folder name already exists. Skipping."
+    fi
     any_files_switch=1
   else
     echo "The file ${f} doesn't look like a Rise 360 SCORM file. Check the filename contains the word 'raw' and unzip manually if needs be."
